Add initial frontend mockup and some simple autocomplete functionality.
I wound up using typeahead.js for the autocomplete. It has an engine for
remote fetching but I haven't spent enough time setting it up yet to
actually pull results. At the moment it just displays programming
languages as tags for suggestions (try 'haskell'). I've only tested this
in chrome and in firefox. I'm not a webapp superstar or anything so I'll
still need to figure out how to set things up so I can make ajax
requests to localhost for testing.

Lastly, there's some simple jQuery stuff for hiding the title when the
user starts typing and to restore it when they click anywhere
(unfortunately, this also means the suggestion area at the moment).
Still, It's a nice visual touch and could help get things cleaned up for
actual results when we decide to have them on the same page (say,
removing the centered search and fading in another on the upper left,
for example if we'd like)