Create template and static directory and route index.
Initial landing page shows up when accessing root ('/') now. Will need
to tie in the form as part of the jinja2 templating and also create a
base template for the search results page. For the time being, we'll
probably just stick to results being on a separate page for ease.