Get unpretty results back from search submissions.
It's not beautiful, but it works. It shows search results in another
page, as well as maintaining the origin externa; 'q' request api. The
two are separated via the distinction in post vs get. I wouldn't
overload it any further. I'm going to clean things up and maybe add a
dictinary to get word language back; but that's not essential and might
be a waste of time.