html templating changes to work towards a reusable search bar template
currently the new template, search.html is just used in /info (until it can be reformatted).
The template is pretty ugly for now, I don't have the patience to deal with the CSS and HTML
right at the moment.

Note that rendering pages now requires a body_class from now on, and body tags should not
be used in templates any more (they have been added to base.html and search.html)